+++ src/data/num/stumpfu.rs
@@ -152,7 +152,8 @@ pub fn to_church() -> Term {
 
 /// Applied to a Stump-Fu-encoded number it produces the equivalent Scott-encoded number.
 ///
-/// TO_SCOTT ≡ λn.n (λm.m SUCC ZERO) (n TRUE n) ≡ λ 1 (λ 1 SUCC ZERO) (1 TRUE 1)
+/// TO_SCOTT ≡ λn.n (λm.m SCOTT_SUCC SCOTT_ZERO) (n TRUE n)
+///          ≡ λ 1 (λ 1 SCOTT_SUCC SCOTT_ZERO) (1 TRUE 1)
 ///
 /// # Example
 /// ```
@@ -165,7 +166,11 @@ pub fn to_church() -> Term {
 /// ```
 pub fn to_scott() -> Term {
     abs(app(
-        abs(app!(Var(1), scott::succ(), scott::zero())),
+        abs(app!(
+            Var(1),
+            scott::succ(),
+            scott::zero()
+        )),
         app!(
             Var(1),
             tru(),
@@ -176,7 +181,8 @@ pub fn to_scott() -> Term {
 
 /// Applied to a Stump-Fu-encoded number it produces the equivalent Parigot-encoded number.
 ///
-/// TO_PARIGOT ≡ λn.n (λm.m SUCC ZERO) (n TRUE n) ≡ λ 1 (λ 1 SUCC ZERO) (1 TRUE 1)
+/// TO_PARIGOT ≡ λn.n (λm.m PARIGOT_SUCC PARIGOT_ZERO) (n TRUE n)
+///            ≡ λ 1 (λ 1 PARIGOT_SUCC PARIGOT_ZERO) (1 TRUE 1)
 ///
 /// # Example
 /// ```
@@ -189,7 +195,11 @@ pub fn to_scott() -> Term {
 /// ```
 pub fn to_parigot() -> Term {
     abs(app(
-        abs(app!(Var(1), parigot::succ(), parigot::zero())),
+        abs(app!(
+            Var(1),
+            parigot::succ(),
+            parigot::zero()
+        )),
         app!(
             Var(1),
             tru(),